chore: migrate "add-action-binary-file" test to jest
