+++ jest-tests/add-action-binary-file-mock/plop-logo.png
